requirements.txt
anthropic>=0.18.0
pillow>=10.0.0
playwright>=1.40.0
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.0.0
